requirements.txt
altair==4.2.2
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boltons==21.0.0
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
cheroot==9.0.0
click==8.1.7
click-spinner==0.1.10
colorlog==6.7.0
comm==0.1.4
dacite==1.8.1
datacommons==1.4.3
datacommons-pandas==0.0.3
datapane==0.16.7
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.4.2
dominate==2.8.0
dulwich==0.21.5
email-validator==2.0.0.post2
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
face==20.1.1
fastapi==0.103.0
fastjsonschema==2.18.0
furl==2.1.3
gitdb==4.0.10
GitPython==3.1.32
glom==23.3.0
gotrue==1.0.4
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.13.0
ipykernel==6.25.1
ipynbname==2023.2.0.0
ipython==8.14.0
itsdangerous==2.1.2
jaraco.functools==3.9.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
micawber==0.5.5
mistune==3.0.1
monotonic==1.6
more-itertools==10.1.0
multimethod==1.9.1
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
numpy==1.25.2
orderedmultidict==1.0.1
orjson==3.9.5
packaging==23.1
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
postgrest==0.10.8
posthog==2.5.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pydantic==1.10.12
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyngrok==5.2.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.1
realtime==1.0.0
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
stack-data==0.6.2
starlette==0.27.0
storage3==0.5.4
streamlit==1.26.0
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
tabulate==0.9.0
tenacity==8.2.3
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.21.2
vega-datasets==0.9.0
watchfiles==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
zipp==3.16.2
